+++ src/query/__tests__/machine-with-data.test.ts
@@ -43,6 +43,7 @@ function makeRefreshing(): TestMachine<TestArgs, TestData> {
         error: null,
         updatedAt: 1000,
         patchState: null,
+        isRetrying: false,
     });
 }