+++ src/query/core/machine/Machine.ts
@@ -30,6 +30,7 @@ export namespace Machine {
             data: null,
             error: null,
             updatedAt: null,
+            isRetrying: false,
         };
         return new MachinePending<TArgs, TData>(state);
     }
@@ -46,6 +47,7 @@ export namespace Machine {
                 error: null,
                 updatedAt: snapshot.updatedAt,
                 patchState: null,
+                isRetrying: false,
             };
             return new MachineRefreshing<TArgs, TData>(state);
         }