+++ src/query/core/machine/MachineSuccess.ts
@@ -61,6 +61,7 @@ export class MachineSuccess<TArgs, TData> extends MachineWithData<TArgs, TData> 
             error: null,
             updatedAt: this.state.updatedAt,
             patchState: this.state.patchState,
+            isRetrying: false,
         };
         return new MachineRefreshing<TArgs, TData>(state);
     }